Move InvalidQueryException to error_classes
